requirements.txt
aniso8601==7.0.0
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
autopep8==1.4.4
Babel==2.12.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
comtypes==1.1.7
contourpy==1.0.7
cryptography==36.0.2
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==2.2
django-allauth==0.50.0
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-countries==5.3.3
django-crispy-forms==1.14.0
django-database-view==0.3.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-jsonfield==1.4.1
django-phonenumber-field==5.2.0
django-tables2==2.4.1
django-taggit==3.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.6
et-xmlfile==1.1.0
filelock==3.6.0
fonttools==4.33.3
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-resources==5.12.0
Jinja2==3.1.2
Js2Py==0.70
kiwisolver==1.4.4
lxml==4.8.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.0
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.9
oscrypto==1.3.0
packaging==20.4
pdfkit==1.0.0
pep8==1.7.1
phonenumbers==8.12.34
Pillow==9.1.0
pipwin==0.5.0
platformdirs==2.5.2
playsound==1.2.2
promise==2.3
psycopg2==2.9.1
PyAudio @ file:///C:/Users/Admin/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycodestyle==2.5.0
pycparser==2.21
pydyf==0.2.0
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyjsparser==2.7.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF3==1.0.6
pyphen==0.12.0
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
python-docx==0.8.11
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==228
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stripe==2.27.0
svglib==1.2.1
tablib==3.2.1
text-unidecode==1.3
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
Unipath==1.1
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.14.1
weasyprint==55.0
webencodings==0.5.1
whitenoise==5.3.0
wikipedia==1.4.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
zopfli==0.2.1
